+++ packages/primeng/src/index.ts
@@ -6,3 +6,4 @@ export * from './lib/time-slots.component';
 export * from './lib/markdown-editor.component';
 export * from './lib/captcha.component';
 export * from './lib/image-crop.component';
+export * from './lib/phone-input.component';